claude/enterprise-kb/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9

# 数据库
sqlalchemy==2.0.30
asyncpg==0.29.0
alembic==1.13.1
pgvector==0.2.5

# 鉴权
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# 配置管理
pydantic==2.7.1
pydantic-settings==2.2.1

# AI / LLM
openai==1.30.1
tiktoken==0.7.0

# 文档解析
pymupdf==1.24.3
markdown==3.6

# 可观测性
langfuse==2.36.1
structlog==24.2.0

# 评测
ragas==0.1.9
datasets==2.19.1

# 工具
tenacity==8.3.0
httpx==0.27.0
python-dotenv==1.0.1
